Implement csv->rdf function.
Move the iteration over source tables in the resolved metadata
from csvw.minimal into the csvw namespace. Add a
csv->rdf->destintaion function which allows the caller to provide
the Grafter ITripleWriteable instance to write the result statements
to.

Implement the csv->rdf function called by the tests to write the
results to an in-memory repositiory.