Remove default for inherited lang property.
The specification defines a default value of 'und' for the inherited
lang property, which specifies the language for strings in the described
column. The column lang property is used to set the language for string
cell values within the column, however the CSVW tests do not set the
language on strings where the column lang property is unspecified.

Move the StringLike protocol and use it to set URI property values
as scalars.