Fetch titles for each row.
Fetch row titles for each row according to the rowTitles schema property
and associate them with the row record.

Output a csvw:title statement for each title in the row - use the
semantic value if the cell contains a single element or the string
value if the cell contains a list.