Fix CSVW standard output.
Use the same default subject for all cells within a row without an
aboutURL property.

Explicitly set the type of the csvw:rownum value to xsd:integer.
Integer values are given a type of xsd:int by grafter.

Set the base URI to the action URI in the test definition when loading
test data files.