Write generated CSVW tests to file.
Generate CSVW test suite tests and write them to file rather than
loading generating and loading them dynamically. Move the functions
for loading the test manifest and generating test functions to a new
csvw-test.gen namespace, and move the artifacts used in the tests to
a shared csvw-test.impl namespace.