Merge user metadata and embedded metadata before expansion.
Users can provide metadata which contains a minimal description
of the data which should be combined with the implicit schema
embedded in the tabular data file. For example, the metadata
for CSVW test023 contains a dialect definition but no tableSchema
key, so this must be extracted from the referenced tabular file.

Don't expand metadata during the parsing process, and instead
manually expand it after combining it with the schema embedded
in the tabular file if applicable.