Unset prefixes when writing CSVW to file.
Each call to rdf/add causes any prefixes to be written to the file.
csv->rdf->destination makes multiple calls to rdf/add which causes the
prefixes block to be written multiple times to the destination file.

Explicitly unset the URI prefixes when constructing the RDF serialiser
so no prefixes are written to the output.