Use empty URI template if invalid in the metadata.
The specification states:

If the supplied value of a URI template property is not a string
(e.g. if it is an integer), compliant applications MUST issue a
warning and proceed as if the property had been supplied with an
empty string.

Add a custom method for resolving the expanded template strings
so empty strings do not modify the base URI.