Return all cell errors after writing CSVW statements.
Replace table-statements method with a write-table-statements method which is responsible
for writing the statements for each mode to the destination. Update each mode to
collect all encountered cell errors when processing each input row and return them
to the caller.

Merge any cell errors with the warnings encountered when processing the user metadata.