Validate 'other' datatypes against the format regex if provided.
Split the parsing of cell data based on those with a provided format
property, and those without. Any 'other' datatypes (i.e. not a number,
boolean, date/time or duration) should provide a regex to validate
the raw cell data against. Validate the cell data against the format
regex and output a cell error if it does not match.