Log warning if located metadata does not reference tabular file.
A number of the CSVW tests expect a warning to be logged if the metadata
location process encounters a metadata file which does not contain
a reference to the tabular file. Add logging namespace along with
an in-memory logger implementation which accumulates log messages.
Add any encountered warning messages to the warnings collection
returned by csvw->rdf.