Validate date/time and duration values according to datatype bounds.
Parse any minimum or maximum values specified in a datatype
definition according to the base datatype.

Change the implementations of parse for date and time datatypes to
return the same types as the corresponding parse-format methods
rather than XMLGregorianCalendar.

Add protocol for comparing ordered cell datatypes. Use it to validate
any bounds on a datatype definition are valid.