Forbid integer or decimal groups in formats that don't specify them.
If a numeric format does not specify the group size for either the
integer or decimal parts, do not allow groups in those parts when
parsing numbers e.g. the format ##0 does not specify a primary or
secondary group, so the string '1,234' should be rejected since it
contains a group.