Add command-line interface.
Add main function which implements a command-line interface for the CSVW
process. Users must specify either a tabular or user metadata source to
start from. The output is written to standard output by default but a
destination file can be specified. If a destination file is provided,
output format will be inferred from the file extension if possible,
otherwise the format will be set to Turtle. Users can specify the CSVW
mode (standard or minimal) to run using the -m option. The default
mode is standard.