Store calculated column name under namespaced key.
The original column name is needed when validating schemas, but metadata
parse time is the best(?) place to calculate the resulting column name
since the context is available to get the current language. Store the
calculated column name for a column under the ::name key. Add
accessor function for the column name and use it from other namespaces.