Create iterator wrapper which takes ownership of an underlying source.
Create owning-iterator function which takes ownership of a closeable
resource which is used to construct an inner iterator. The owning
iterator lazily creates the inner iterator, and closes the resource
when the inner iterator is exhausted, or if an exception is thrown
during the iteration.

Modify read-rows to take a tabular source instead of a Reader instance.
read-rows now constructs the reader and passes the ownership to the
returned sequence of source records. This means clients no longer
need to manually close the reader, but the resulting sequence must
eventually be iterated over to completion to ensure the contained
reader is eventually closed.