Return a sequence from csv->rdf.
Change the interface of csv->rdf so it now returns a single lazy
sequence of statements. The returned sequence contains the lazy
concatenation of the statements for each unsuppressed table in the
metadata. The returned sequence should be completely enumerated
to ensure all resources are closed.

Move logging of cell errors into annotate-row rather than doing it
separately within the CSVW modes as the parsed rows are processed.