Add protocol for loading JSON documents.
Add JSONSource protocol for loading a JSON structure from a given
source. Add implementations for URI and File which load the document
using util/read-json.

Add MapMetadataSource record containing a URI and metadata document
map. This implements URIable and JSONSource and allows user metadata
to be loaded from in-memory maps without requiring them to be
serialised and deserialised.