Create protocol for resolving tabular metadata.
Replace the metadata/get-metadata multimethod with a method on a new
MetadataLocator protocol. Implement the protocol for File and URI,
where the URI implementation behaves similarly to the old method.

The File implementation does not need to consider the metadata location
process defined in the specification and returns the embedded metadata
immediately.