Add RowSource protocol.
Add RowSource protocol which views tabular data sources as a
read-once sequence of row records. This allows data sources
to be created which differ from the underlying data they are
derived from.

During the CSVW process, tabular sources are resovled by their URI
in the metadata. Add a :table-resolver option which allows callers
to manage the resolution of a tabular URI into the corresponding
RowSource instance.