Consider languages compatible if one is a prefix of another.
Language codes should be considered compatible if a truncation of the
longer tag defined by the BPC-47 matches the shorter tag. In this case
the shortest tag should be taken as the language when finding compatible
titles. Update intersect-titles to search for the shortest common
language tag when searching for common titles.

Add bcp47 namespace for parsing language tags according to the BCP-47
specification. Add functions for finding all the truncations of a
tag, and use it to find matching truncations when validating column
compatibility.