Generate EARL report for CSVW test cases.
Issue #13 - Add csv2rdf.earl namespace for generating EARL reports
for the CSVW test cases. write-csvw-earl-report overrides the
clojure.test report function with an implementation that writes
RDF for the test summary to a bound RDF serialiser.

Update csvw-test.gen to add the test name and comment to the
metadata for each test case and include these in the test report
output.