Add eager version of split-at.
Issue #29 - Add eager version of split-at which consumes the first n
items in the input sequence and returns a pair of the consumed
prefix and a lazy sequence for the remainder. Use this function
within csv/get-header and csv/rows->embedded-metadata.

split-at returns two lazy sub-sequences of the input sequence. This
caused a space leak since the data rows of the input file were being
processed before the lazy sequences for the comment and header rows
were fully evaluated. This resulted in the entire input row sequence
being held in memory since the head of the sequence was still required.