Read JSON body from URIs with source/get-json.
Issue #41 - Read the body of the HTTP response from GETting a URI
source using get-source. The default response type for GET requests
in the underlying clj-http library is String but this may be changed
in future.

Add source-test namespace and add test for the parsing behaviour.
Move the definition of TestHttpClient into test-common.