Exit with a non-zero exit code on an exception in main
Issue #49 - Wrap main in a try/catch block and exit with a non-zero
exit code in the event of an exception being thrown. The JVM does
not guarantee to set a non-zero code if the program terminates due
to an unhandled exception.

Add a display-error method for displaying unhandled exceptions based
on the :type key of the associated ex-data map if one exists.

Refactor the CLI argument parser to raise an exception if the
arguments are invalid. Display the usage summary in the handler for
invalid CLI exceptions.