Fix logging configuration
Issue #322 - Warnings are not being displayed because the log
configuration no longer works.

The SLF44 API was updated to version 2, which now uses the service
provider API to locate logging implementations. Bindings which target
the 1.7 API are ignored.

Change the log4j binding implementation to log4j-slf4j2-impl which
target version 2 of the SLF4J API.

Update the build-app task in build.clj to include the resources of
the with-logging profile which includes the log4j2.xml configuration
file.